requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai==0.28.1
openai-whisper @ git+https://github.com/openai/whisper.git@b38a1f20f4b23f3f3099af2c3e0ca95627276ddf
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
pooch==1.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.3.3
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
yarl==1.9.2
zipp==3.17.0
